Modificación de modelo pasó a controlador
Se pasaron dos clases que estaban en el modelo que debían estar en
controlador (com.banco.controlador.archivos/hilos)
