+++ nbproject/private/private.properties
@@ -1,2 +1,2 @@
 compile.on.save=true
-user.properties.file=C:\\Users\\Luis Enrique\\AppData\\Roaming\\NetBeans\\8.2\\build.properties
+user.properties.file=/Users/juanjoseadanlopez/Library/Application Support/NetBeans/8.2/build.properties
